social_scheduler_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
apscheduler
pydantic
pydantic-settings
requests
httpx
python-dotenv
cryptography